Fix circular import bug
